requirements.txt
natsort==8.2.0
numpy==1.23.5
Pillow==9.3.0
torch==1.13.0
tqdm==4.64.1
transformers==4.25.1
timm == 0.4.9
git+https://github.com/openai/CLIP.git
scipy~=1.9.3
torchvision~=0.14.0